requirements.txt
-e .

aws_cdk.aws_eks